fix some problems & update drawing code
- use 4 vertex instead 6 for drawing sphere
- reorganize font module & use GL_TRIANGLE_FAN for drawing
- check nullable callback functions in window module

+++ src/font.hpp
@@ -17,7 +17,7 @@ public:
     ~gFont();
 private:
     const uint16_t FONT_LIST_SIZE = 256;
-    void make_dlist( FT_Face face, int ch );
+    void make_dlist( FT_Face face, uint16_t ch );
     GLuint * tex;
     GLuint list;
 };
